результат деления целыми числами, вывод результата в int

+++ src/Delimetr.java
@@ -1,7 +1,7 @@
 
 public class Delimetr extends Computation {
 
-    public float showResult(int a, int b){
+    public int showResult(int a, int b){
         if (a<1||a>10||b<1||b>10){throw new ArithmeticException("параметры введены не в допуске условия задачи.");}
         return a/b;
     }
